Refactor transaction assembly strategies by moving them  to a new package, enhancing organization and maintainability.
