+++ src/main/java/info/mackiewicz/bankapp/transaction/service/assembler/strategies/OwnTransferAssemblyStrategy.java
@@ -1,9 +1,10 @@
-package info.mackiewicz.bankapp.transaction.service.assembler;
+package info.mackiewicz.bankapp.transaction.service.assembler.strategies;
 
 import info.mackiewicz.bankapp.account.model.Account;
 import info.mackiewicz.bankapp.account.service.AccountService;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.OwnTransferRequest;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
+import info.mackiewicz.bankapp.transaction.service.assembler.TransactionAssemblyStrategy;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
